eCommerce-backend/src/main/java/com/app/domain/item/services/CategoryService.java
package com.app.domain.item.services;


import com.app.domain.item.entities.Category;
import com.app.domain.item.repositories.CategoryRepository;
import jakarta.transaction.Transactional;
import jakarta.validation.Valid;
import org.springframework.stereotype.Service;

@Service
public class CategoryService {

    private final CategoryRepository categoryRepository;

    public CategoryService(CategoryRepository categoryRepository) {
        this.categoryRepository = categoryRepository;
    }

    @Transactional
    public Category save(Category category) {
        return categoryRepository.save(category);
    }

    @Transactional
    public Category update(Category category) {
        return save(category);
    }

}
